Separate out triage and comparison implementation
